move some utils to common
