+++ common/Cargo.toml
@@ -15,6 +15,7 @@ foundry-config = { path = "../config" }
 
 # eth
 ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
+ethers-solc = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
 # cli
 clap = { version = "3.0.10", features = [
     "derive",
@@ -27,3 +28,4 @@ clap = { version = "3.0.10", features = [
 serde = "1.0.133"
 serde_json = "1.0.67"
 thiserror = "1.0.31"
+eyre = "0.6"
