+++ common/src/lib.rs
@@ -4,10 +4,12 @@
 
 pub mod calc;
 pub mod constants;
+pub mod contracts;
 pub mod errors;
 pub mod evm;
 pub mod fmt;
 pub mod fs;
 pub mod traits;
 pub use constants::*;
+pub use contracts::*;
 pub use traits::*;
