+++ evm/src/fuzz/mod.rs
@@ -7,8 +7,7 @@ use ethers::{
     abi::{Abi, Function, Token},
     types::{Address, Bytes, Log},
 };
-use foundry_common::calc;
-use foundry_utils::types::ContractsByAddress;
+use foundry_common::{calc, contracts::ContractsByAddress};
 pub use proptest::test_runner::{Config as FuzzConfig, Reason};
 use proptest::test_runner::{TestCaseError, TestError, TestRunner};
 use serde::{Deserialize, Serialize};
